--- frontend/lib/feature/categories/bloc/category_event.dart
@@ -1,10 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class CategoryEvent extends Equatable {
-  const CategoryEvent();
-
-  @override
-  List<Object> get props => [];
-}
-
-class LoadProducts extends CategoryEvent {}
