--- frontend/lib/feature/categories/bloc/category_state.dart
@@ -1,17 +1,0 @@
-import '../models/category_model.dart';
-
-abstract class CategoryState {}
-
-class CategoryInitial extends CategoryState {}
-
-class CategoryLoading extends CategoryState {}
-
-class CategoryLoaded extends CategoryState {
-  final List<CategoryModel> categories;
-  CategoryLoaded(this.categories);
-}
-
-class CategoryError extends CategoryState {
-  final String message;
-  CategoryError(this.message);
-}
