--- frontend/lib/feature/home/bloc/home_event.dart
@@ -1,3 +1,0 @@
-abstract class HomeEvent {}
-
-class LoadCategories extends HomeEvent {}
