+++ frontend/lib/feature/home/bloc/product_bloc.dart
@@ -1,0 +1,23 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:equatable/equatable.dart';
+import '../../../models/product_model.dart';
+import '../repositories/product_repository.dart';
+
+part 'product_event.dart';
+part 'product_state.dart';
+
+class ProductBloc extends Bloc<ProductEvent, ProductState> {
+  final ProductRepository repository;
+
+  ProductBloc(this.repository) : super(ProductLoading()) {
+    on<LoadProducts>((event, emit) async {
+      emit(ProductLoading());
+      try {
+        final products = await repository.fetchProducts();
+        emit(ProductLoaded(products));
+      } catch (e) {
+        emit(const ProductError("Error al cargar productos"));
+      }
+    });
+  }
+}
