+++ frontend/lib/feature/home/bloc/product_event.dart
@@ -1,0 +1,10 @@
+part of 'product_bloc.dart';
+
+abstract class ProductEvent extends Equatable {
+  const ProductEvent();
+
+  @override
+  List<Object> get props => [];
+}
+
+class LoadProducts extends ProductEvent {}
