+++ frontend/lib/feature/home/bloc/product_state.dart
@@ -1,0 +1,28 @@
+part of 'product_bloc.dart';
+
+abstract class ProductState extends Equatable {
+  const ProductState();
+
+  @override
+  List<Object> get props => [];
+}
+
+class ProductLoading extends ProductState {}
+
+class ProductLoaded extends ProductState {
+  final List<Product> products;
+
+  const ProductLoaded(this.products);
+
+  @override
+  List<Object> get props => [products];
+}
+
+class ProductError extends ProductState {
+  final String message;
+
+  const ProductError(this.message);
+
+  @override
+  List<Object> get props => [message];
+}
