--- frontend/lib/feature/home/models/categories_model.dart
@@ -1,7 +1,0 @@
-class Categories {
-  final String name;
-  final String description;
-  final double precio;
-
-  Categories({required this.name, required this.description, required this.precio});
-}
