+++ frontend/lib/feature/home/repositories/product_repository.dart
@@ -1,12 +1,12 @@
 import 'dart:convert';
+import 'package:design_alma/global/api.dart';
 import 'package:http/http.dart' as http;
-import '../models/product_model.dart';
 
-class ProductService {
-  static const String _baseUrl = 'http://localhost:3000';
+import '../../../models/product_model.dart';
 
-  static Future<List<Product>> fetchProducts() async {
-    final response = await http.get(Uri.parse('$_baseUrl/api/products?locale=es'));
+class ProductRepository {
+  Future<List<Product>> fetchProducts() async {
+    final response = await http.get(Uri.parse('${Api.baseUrl}/api/products'));
 
     if (response.statusCode == 200) {
       final List<dynamic> data = jsonDecode(response.body);
