+++ frontend/lib/feature/home/views/home_error_view.dart
@@ -2,28 +2,13 @@ import 'package:flutter/material.dart';
 
 class HomeErrorView extends StatelessWidget {
   final String message;
-  final VoidCallback onRetry;
 
-  const HomeErrorView({
-    super.key,
-    required this.message,
-    required this.onRetry,
-  });
+  const HomeErrorView({super.key, required this.message});
 
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(message, style: const TextStyle(color: Colors.red)),
-          const SizedBox(height: 12),
-          ElevatedButton(
-            onPressed: onRetry,
-            child: const Text("Reintentar"),
-          ),
-        ],
-      ),
+      child: Text(message),
     );
   }
 }
