+++ frontend/lib/feature/home/views/home_loading_view.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 
 class HomeLoadingView extends StatelessWidget {
-  const HomeLoadingView({
-    super.key,
-  });
+  const HomeLoadingView({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Center(child: CircularProgressIndicator());
+    return const Center(
+      child: CircularProgressIndicator(),
+    );
   }
 }
