+++ frontend/lib/global/api.dart
@@ -1,0 +1,3 @@
+class Api {
+  static const String baseUrl = 'http://localhost:3000';
+}
