+++ frontend/lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:design_alma/routes/routes.dart';
-import 'package:design_alma/screens/logo_intro.dart'; // 👈 Tu animación Lottie
+import 'package:design_alma/screens/logo_intro.dart';
 
-void main() {
+void main() async {
   runApp(const MyApp());
 }
 
@@ -16,7 +16,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'DesignAlma App',
-      home: const LogoIntro(), // 👈 Se muestra enseguida al abrir
+      home: const LogoIntro(),
       routes: appRoute.routes,
     );
   }
