+++ frontend/lib/widgets/custom_appbar.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class CustomAppBar extends StatelessWidget {
+  const CustomAppBar({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        // Logo
+        Image.asset(
+          'assets/logo.png',
+          height: 100,
+          color: Colors.black,
+        ),
+        Row(
+          children: [
+            IconButton(
+              icon: const Icon(Icons.search),
+              onPressed: () {},
+            ),
+            IconButton(
+              icon: const Icon(Icons.more_vert_outlined),
+              onPressed: () {},
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
