+++ frontend/pubspec.yaml
@@ -29,3 +29,4 @@ flutter:
   assets:
     - assets/animations/logo_intro.json
     - assets/logo.png
+    - assets/icon_trending.png
